#173 moved default_duty_agenda to filehub class
