+++ src/real/admin_change.py
@@ -1,18 +1,6 @@
-from src._instrument.file import delete_dir, get_integer_filenames
+from src._instrument.file import get_integer_filenames
 from src.change.filehub import FileHub
 from src.agenda.agenda import AgendaUnit
-
-
-class ChangeFileMissingException(Exception):
-    pass
-
-
-def create_save_changeunit(
-    x_filehub: FileHub, before_agenda: AgendaUnit, after_agenda: AgendaUnit
-):
-    new_changeunit = x_filehub._create_new_changeunit()
-    new_changeunit._bookunit.add_all_different_agendaatoms(before_agenda, after_agenda)
-    x_filehub.save_change_file(new_changeunit)
 
 
 def _merge_changes_into_agenda(x_filehub: FileHub, x_agenda: AgendaUnit) -> AgendaUnit:
